latticelib: new fn subtractMeanFromEnergyMatrix
I found that the energy matrix reading function in latticelib adds 3.16 (subtracts -3.16) from every entry because that is the mean value in MJ96 Table 3 upper triangle. Well this is not compatible with reading different energy tables, so I got rid of this, and I added a new function, subtractMeanFromEnergyMatrix, in case it's still desired to subtract away the mean.

Code appearance has been cleaned up where I've had to read it (line breaks and indentation).