0.0.9: new fitness function for folding_evolution
We calculate fitness in a revised way. Using latpack/1.9.1-13 which makes a calculation for us when we run the simulation. Protein has a probability to degrade that increases if it's not in the native state. Integral of this varying probability with whether the protein is in the native state over the simulation is now the fitness estimated from one trajectory.