NonsynonymousOnly mode samples all AAs uniformly
Previously, NonsynonymousOnly mode would make nucleotide changes until the codon change is nonsynonymous. Now with a new function implemented in the gencode.c library, `AAMutateNucSequence`, amino acid changes are sampled uniformly.

Also, I noticed a problem with the previous version of the mutation code in `folding_evolution`. The mutation function is part of a while loop, and that means that for each iteration of the loop, a nucleotide change would be made. I was running simulations with MutateAll mode, and typically, the loop only iterated once, so I guess I never noticed. But the previous mode of operation would be problematic for SynonymousOnly and also for cases in which a mutation to a stop codon occurred.