0.0.14-draft1: Adaptive posttrans; fitness func
We want to explore degradation timescale, so it makes sense for posttranslation time to change with degradation such that the posttranslation period ends when the protein has a 6.25% chance of surviving if it remained unfolded for the entirety of the posttranslation period.

Additionally, the fitness function sharply rising with pnat is not sensible. We filter protein output through x/(x+x0), with x0 = 0.01. (it's possible this constant should be commandline adjustable)