Simulation mode for synonymous mutations.
The effect of synonymous mutations is to slow translation speed for specific residues. This version of the simulation program only allows synonymous mutations to occur. There is an option to stop the simulation as soon as a single synonymous mutation is accepted.

This version was branched off of aa892fb (new_fitness_eval).